Login
Login page has been updaed to show login error. Changed the login view.

+++ aps/php/RegisterNewUser.php
@@ -1,41 +1,52 @@
 <?php
   include 'dbconfig.php';
 
-  $mysqli = mysqli_connect($dbhost, $dbuser, $dbpass, $dbname);
-  
   if (!mysqli_connect_errno()) {
-    $username = $_GET['username'];
-    $firstname = $_GET['firstname'];
-    $lastname = $_GET['lastname'];
-    $email = $_GET['email'];
-    $phone = $_GET['phone'];
-    $street = $_GET['street'];
-    $city = $_GET['city'];
-    $state = $_GET['state'];
-    $zip = $_GET['zip'];
-    $password = $_GET['password'];
+    $username = trim($_GET['username']);
+    $firstname = trim($_GET['firstname']);
+    $lastname = trim($_GET['lastname']);
+    $email = trim($_GET['email']);
+    $phone = trim($_GET['phone']);
+    $street = trim($_GET['street']);
+    $city = trim($_GET['city']);
+    $state = trim($_GET['state']);
+    $zip = trim($_GET['zip']);
+    $password = trim($_GET['password']);
     $hashed_password = password_hash($password, PASSWORD_DEFAULT);
 
-    $result = mysqli_query($mysqli, "INSERT INTO `Customer` (`Username`, `Fname`, `Lname`, `Street`, `City`, `State`, `Zipcode`, `Phone`, `Email`) VALUES ('$username','$firstname','$lastname', '$street', '$city', '$state', '$zip', '$phone', '$email')");
+    if(empty($username) || empty($firstname) || empty($lastname) || empty($email) || empty($street) || empty($city) || empty($state) || empty($zip) || empty($password)) {
+      echo "{\"records\":[{\"Status\":\"FAIL\"}]}";
+      exit();
+    }
+    else {
+      $mysqli = mysqli_connect($dbhost, $dbuser, $dbpass, $dbname);
 
-    //error_log("INSERT INTO `Customer` (`Username`, `Fname`, `Lname`, `Street`, `City`, `State`, `Zipcode`, `Phone`, `Email`) VALUES ('$username','$firstname','$lastname', '$street', '$city', '$state', '$zip', '$phone', '$email')");
+      $result = mysqli_query($mysqli, "INSERT INTO `Customer` (`Username`, `Fname`, `Lname`, `Street`, `City`, `State`, `Zipcode`, `Phone`, `Email`) VALUES ('$username','$firstname','$lastname', '$street', '$city', '$state', '$zip', '$phone', '$email')");
+      //error_log("INSERT INTO `Customer` (`Username`, `Fname`, `Lname`, `Street`, `City`, `State`, `Zipcode`, `Phone`, `Email`) VALUES ('$username','$firstname','$lastname', '$street', '$city', '$state', '$zip', '$phone', '$email')");
 
-    if($result) {
-      $userresult = mysqli_query($mysqli, "INSERT INTO `users` (`Username`, `Password`) VALUES ('$username', '$hashed_password')");
+      if($result) {
+        $userresult = mysqli_query($mysqli, "INSERT INTO `users` (`Username`, `Password`) VALUES ('$username', '$hashed_password')");
 
-      if($userresult) {
-        mysqli_close($mysqli);
+        if($userresult) {
+          mysqli_close($mysqli);
 
-        echo "{\"records\":[{\"Status\":\"SUCCESS\"}]}";
-        exit();
-       }
-       else {
-        mysqli_query($mysqli, "DELETE FROM `Customer` WHERE `Username` = '$username')");
+          echo "{\"records\":[{\"Status\":\"SUCCESS\"}]}";
+          exit();
+         }
+         else {
+          mysqli_query($mysqli, "DELETE FROM `Customer` WHERE `Username` = '$username')");
+          mysqli_close($mysqli);
+
+          echo "{\"records\":[{\"Status\":\"FAIL\"}]}";
+          exit();
+         }
+      }
+      else {
         mysqli_close($mysqli);
 
         echo "{\"records\":[{\"Status\":\"FAIL\"}]}";
         exit();
-       }
+      }
     }
     else {
       mysqli_close($mysqli);
@@ -44,10 +55,4 @@
       exit();
     }
   }
-  else {
-      mysqli_close($mysqli);
-
-      echo "{\"records\":[{\"Status\":\"FAIL\"}]}";
-      exit();
-    }
 ?>
